backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.5
google-auth-oauthlib==0.4.6
google-api-python-client==2.86.0 